\name{seedDisp-package}
\alias{seedDisp-package}
\alias{seedDisp}
\docType{package}
\title{Seed dispersal routines }
\description{  Disperses seeds by using different routines }
\details{
\tabular{ll}{
Package: \tab seedDisp\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2010-12-08\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

There is only one function in this package, windDisp
}
\author{
Rainer M Krug \email{Rainer@krugs.de}
}
\keyword{ package }
