% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{seedDisp-package}
\alias{seedDisp-package}
\title{Disperses seeds by using different routines}
\description{
This package is a collection of different seed sispersal routines.
}
\details{
\tabular{ll}{
Package: \tab seedDisp\cr
Type: \tab Package\cr
Version: \tab 0.6-0\cr
Date: \tab 2016-01-19_14-15\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Rainer M Krug \email{Rainer@krugs.de}
}

