\docType{package}
\name{seedDisp-package}
\alias{seedDisp-package}
\title{Disperses seeds by using different routines}
\description{
  This package is a collection of different seed sispersal
  routines.
}
\details{
  ##' \tabular{ll}{ Package: \tab seedDisp\cr Type: \tab
  Package\cr Version: \tab 0.4.13\cr Date: \tab
  2013-09-09_14-51\cr License: \tab GPL (>= 2)\cr LazyLoad:
  \tab yes\cr }
}
\author{
  Rainer M Krug \email{Rainer@krugs.de}
}

